+++ Source/Testably.Expectations/That/Collections/ThatCollectionShould.All.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using Testably.Expectations.Core;
 using Testably.Expectations.Options;
+using Testably.Expectations.That.Collections;
 
 // ReSharper disable once CheckNamespace
 namespace Testably.Expectations;
