+++ Source/Testably.Expectations/That/Collections/ThatCollectionShould.AtMost.cs
@@ -2,6 +2,7 @@
 using System.Runtime.CompilerServices;
 using Testably.Expectations.Core;
 using Testably.Expectations.Options;
+using Testably.Expectations.That.Collections;
 
 // ReSharper disable once CheckNamespace
 namespace Testably.Expectations;
