+++ Source/Testably.Expectations/That/Collections/ThatCollectionShould.Between.cs
@@ -3,6 +3,7 @@ using System.Runtime.CompilerServices;
 using Testably.Expectations.Core;
 using Testably.Expectations.Options;
 using Testably.Expectations.Results;
+using Testably.Expectations.That.Collections;
 
 // ReSharper disable once CheckNamespace
 namespace Testably.Expectations;
