+++ Source/Testably.Expectations/That/Collections/ThatEnumerableShould.AtLeast.cs
@@ -2,6 +2,7 @@
 using System.Runtime.CompilerServices;
 using Testably.Expectations.Core;
 using Testably.Expectations.Options;
+using Testably.Expectations.That.Collections;
 
 // ReSharper disable once CheckNamespace
 namespace Testably.Expectations;
@@ -16,7 +17,6 @@ public static partial class ThatEnumerableShould
 		int minimum, [CallerArgumentExpression("minimum")] string doNotPopulateThisValue = "")
 	{
 		source.ExpectationBuilder.AppendMethodStatement(nameof(AtLeast), doNotPopulateThisValue);
-		return new QuantifiableCollection<TItem, IEnumerable<TItem>>(source,
-			CollectionQuantifier.AtLeast(minimum));
+		return new(source, CollectionQuantifier.AtLeast(minimum));
 	}
 }
