+++ Source/Testably.Expectations/That/Collections/ThatQuantifiableCollectionShould.Be.cs
@@ -6,6 +6,7 @@ using Testably.Expectations.Core.Constraints;
 using Testably.Expectations.Core.EvaluationContext;
 using Testably.Expectations.Formatting;
 using Testably.Expectations.Results;
+using Testably.Expectations.That.Collections;
 
 // ReSharper disable once CheckNamespace
 namespace Testably.Expectations;
