+++ Source/Testably.Expectations/That/Collections/ThatQuantifiableCollectionShould.BeEquivalentTo.cs
@@ -7,6 +7,7 @@ using Testably.Expectations.Core.Equivalency;
 using Testably.Expectations.Core.EvaluationContext;
 using Testably.Expectations.Options;
 using Testably.Expectations.Results;
+using Testably.Expectations.That.Collections;
 
 // ReSharper disable once CheckNamespace
 namespace Testably.Expectations;
